\name{SDD-package}
\alias{SDD-package}
\docType{package}
\title{SDD - Serial Dependence Diagrams.}
\description{
This package allows for Serial Dependence Diagrams applicable to both linear and nonlinear
time series.
}
\details{
\tabular{ll}{
Package: \tab SDD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-11\cr
License: \tab GNU-2\cr
}
}
\references{
Bagnato L, De Capitani L, Punzo A (2013a). Detecting Serial Dependencies with the Reproducibility Probability Autodependogram. \emph{Advances in Statistical Analysis}. doi:10.1007/s10182-013-0208-y.

Bagnato L, De Capitani L, Punzo A (2013b). Testing Serial Independence via Density-Based Measures of Divergence.\emph{Methodology and Computing in Applied Probability}. doi:10.1007/s11009-013-9320-4.

Bagnato L, Punzo A, Nicolis O (2012). The autodependogram: a graphical device to investigate serial dependences.\emph{Journal of Time Series Analysis}, \bold{33}(2), 233-254.

Bagnato L, Punzo A (2013). Using the Autodependogram in Model Diagnostic Checking. In N Torelli, F Pesarin, A Bar-Hen (eds.), \emph{Advances in Theoretical and Applied Statistics, Studies in Theoretical and Applied Statistics}, pp. 129-139. Springer-Verlag, Berlin-Heidelberg

}
\author{
Luca Bagnato, Lucio De Capitani, Angelo Mazza and Antonio Punzo

Maintainer: Angelo Mazza <a.mazza@unict.it>
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{ADF}}, \code{\link{plot.SDD}}, \code{\link{SMI}}
}