\name{SDD-package}
\alias{SDD-package}
\docType{package}
\title{SDD - Serial Dependence Diagrams.}
\description{
This package allows for serial dependence diagrams applicable to both linear and nonlinear
time series.
}
\details{
\tabular{ll}{
Package: \tab SDD\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-02-21\cr
License: \tab GNU-2\cr
}
}
\references{
Bagnato L, De Capitani L, Punzo A (2013a). Improving the autodependogram using the Kulback-Leibler divergence. \emph{Communications in Statistics - Simulation and Computation}. (Submitted).

Bagnato L, De Capitani L, Punzo A (2013b). Testing Serial Independence via Density-Based Measures of Divergence. \emph{Methodology and Computing in Applied Probability}. doi:10.1007/s11009-013-9320-4.

Bagnato L, De Capitani L, Punzo A (2014). Detecting Serial Dependencies with the Reproducibility Probability Autodependogram. \emph{Advances in Statistical Analysis}, \bold{98}(1), 35-61.

Bagnato L, Punzo A (2010). On the Use of \eqn{\chi^2} Test to Check Serial Independence. \emph{Statistica & Applicazioni}, \bold{VIII}(1), 57-74.

Bagnato L, Punzo A (2012). Checking Serial  Independence of Residuals from a Nonlinear Model. In W Gaul, A Geyer-Shulz, L Schmidt-Thieme, J Kunze (eds.), \emph{Challenges at the Interface of Data Analysis, Computer Science, and Optimization}, volume XIV of \emph{Studies in Classification, Data Analysis and Knowledge Organization}, pp. 203-211. Springer-Verlag, Berlin Heidelberg.

Bagnato L, Punzo A, Nicolis O (2012). The autodependogram: a graphical device to investigate serial dependencies. \emph{Journal of Time Series Analysis}, \bold{33}(2), 233-254.

Bagnato L, Punzo A (2013). Using the Autodependogram in Model Diagnostic Checking. In N Torelli, F Pesarin, A Bar-Hen (eds.), \emph{Advances in Theoretical and Applied Statistics}, volume XIX of \emph{Studies in Theoretical and Applied Statistics}, pp. 129-139. Springer-Verlag, Berlin Heidelberg.


}
\author{
Luca Bagnato, Lucio De Capitani, Angelo Mazza and Antonio Punzo

Maintainer: Angelo Mazza a.mazza@unict.it
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{ADF}}, \code{\link{plot.SDD}}, \code{\link{SMI}}
}